Pero Kanger Iz Here Bois

+++ Yukki/Plugins/Multi-Assistant/Assistant.py
@@ -1,3 +1,4 @@
+from config import PM_PERMIT
 from inspect import getfullargspec
 
 from pyrogram import Client, filters
@@ -27,35 +28,36 @@ flood = {}
     & ~filters.user(SUDOERS)
 )
 async def awaiting_message(client, message):
-    if await is_on_off(5):
-        try:
-            await client.forward_messages(
-                chat_id=LOG_GROUP_ID,
-                from_chat_id=message.from_user.id,
-                message_ids=message.message_id,
+    if PM_PERMIT == "TRUE":
+        if await is_on_off(5):
+            try:
+                await client.forward_messages(
+                    chat_id=LOG_GROUP_ID,
+                    from_chat_id=message.from_user.id,
+                    message_ids=message.message_id,
+                )
+            except Exception as err:
+                pass
+        user_id = message.from_user.id
+        if await is_pmpermit_approved(user_id):
+            return
+        async for m in client.iter_history(user_id, limit=6):
+            if m.reply_markup:
+                await m.delete()
+        if str(user_id) in flood:
+            flood[str(user_id)] += 1
+        else:
+            flood[str(user_id)] = 1
+        if flood[str(user_id)] > 5:
+            await message.reply_text("Spam Detected. User Blocked")
+            await client.send_message(
+                LOG_GROUP_ID,
+                f"**Spam Detect Block On Assistant**\n\n- **Blocked User:** {message.from_user.mention}\n- **User ID:** {message.from_user.id}",
             )
-        except Exception as err:
-            pass
-    user_id = message.from_user.id
-    if await is_pmpermit_approved(user_id):
-        return
-    async for m in client.iter_history(user_id, limit=6):
-        if m.reply_markup:
-            await m.delete()
-    if str(user_id) in flood:
-        flood[str(user_id)] += 1
-    else:
-        flood[str(user_id)] = 1
-    if flood[str(user_id)] > 5:
-        await message.reply_text("Spam Detected. User Blocked")
-        await client.send_message(
-            LOG_GROUP_ID,
-            f"**Spam Detect Block On Assistant**\n\n- **Blocked User:** {message.from_user.mention}\n- **User ID:** {message.from_user.id}",
+            return await client.block_user(user_id)
+        await message.reply_text(
+            f"Hello, I am {MUSIC_BOT_NAME}'s Assistant.\n\nPlease dont spam here , else you'll get blocked.\nFor more Help start :- @{BOT_USERNAME}"
         )
-        return await client.block_user(user_id)
-    await message.reply_text(
-        f"Hello, I am {MUSIC_BOT_NAME}'s Assistant.\n\nPlease dont spam here , else you'll get blocked.\nFor more Help start :- @{BOT_USERNAME}"
-    )
 
 
 @Client.on_message(
